// packages/rive/src/types.ts
export type UseRiveOptions = {
	useDevicePixelRatio: boolean;
	customDevicePixelRatio: number;
	fitCanvasToArtboardHeight: boolean;
	useOffscreenRenderer: boolean;
	shouldResizeCanvasToContainer: boolean;
	shouldUseIntersectionObserver?: boolean;
};

export type Dimensions = {
	width: number;
	height: number;
};